Added support for 'Apps for Alexa'.
In this commit, we added the ability to punch out to directions to a
restroom when the user is on a device that supports AppLinks.

We support Android and iOS devices and punch out to Google and Apple Maps
respectively.

TBD: Should we just use Google Maps on both devices because Refugee Restrooms
is built atop Google Maps and the directions from Google appear to be
more accurate as a result.